requirements.txt
asgiref==3.11.0
cachetools==5.5.2
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-environ==0.12.0
djangorestframework==3.16.1
google-api-core==2.29.0
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.72.0
httplib2==0.31.2
idna==3.11
isodate==0.7.2
mysqlclient==2.2.7
oauthlib==3.3.1
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyparsing==3.3.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
